Pundit

+++ app/controllers/bookings_controller.rb
@@ -2,12 +2,15 @@ class BookingsController < ApplicationController
   before_action :set_office, only: [:new, :edit, :create]
   def new
     @booking = Booking.new
+    authorize @booking
   end
 
   def create
+
     @booking = Booking.new(booking_params)
     @booking.office_id = @office.id
     @booking.user_id = current_user.id
+    authorize @booking
     @booking.save
     redirect_to booking_path(@booking)
     # link needs to be updated to booking#SHOW
@@ -15,10 +18,12 @@ class BookingsController < ApplicationController
 
   def show #confirmation page
     @booking = Booking.find(params[:id])
+    authorize @booking
     # total price of one booking should be added on show page
   end
 
   def destroy
+    authorize @booking
     booking = Booking.find(params[:id])
     @office = booking.office_id
     booking.destroy
